requirements.txt
flask_bcrypt
flask-jwt-extended
attrs==22.1.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
flasgger==0.9.5
Flask-Admin==1.6.0
Flask-Excel==0.0.7
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
jsonschema==4.7.2
lml==0.1.0
#MarkupSafe==2.1.1
mistune==2.0.4
#numpy==1.24.0
openpyxl==3.0.10
#pandas==1.5.2
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyodbc==4.0.35
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
setuptools==65.6.3
six==1.16.0
SQLAlchemy==1.4.39
texttable==1.6.7
urllib3==1.26.13
Werkzeug==2.1.2
wheel==0.38.4
Flask-Cors=5.0.0